Including the Jekyll data plugin as a theme dependency
Data files aren’t loaded for themes by Jekyll in the same way as assets and includes.

There is a plugin which works around this, including merging in config options, which we now depend on.

This means we can set default config options in this theme, rather than require users to set them. In most cases however they still will as they are site specific.

Whilst not as clean as it should/could be - this does give us more flexibility for the future.