Refactoring the Style Kit
This completely refactors the layouts and includes used by this theme.

They are based on the layouts/includes developed for Pug (within the Style Kit Testbed) and do not (re)implement all the functionality present in the previous layouts/includes (particularly navigation menus).

This will be re-added in future commits.

Currently settings for which container to use etc. are defined in a Jekyll data file, rather than the theme’s config file. This will be reviewed nearer release to ensure users are able to change settings, and not have to learn too much about how this theme works.