Adding Jekyll Menu plugin
This adds support for more flexible, and less ad-hoc, menu structures within a Jekyll site.

Within this theme we define two menu’s - both part of the top navigation. The primary navigation is positioned to the right of the site title, the secondary nav is right aligned and by default only contains a ‘launcher’ to signpost other BAS websites.

Pages are added to menus either through front-matter (if content within the current site) - or through the menu’s data-file.

To ensure consistency, the Gem lockfile is now tracked in this repository.